Remove protocol 1 and 2
This two protocol are not used anymore
