--- driver/src/main/java/org/neo4j/driver/internal/messaging/v2/BoltProtocolV2.java
@@ -1,43 +1,0 @@
-/*
- * Copyright (c) 2002-2020 "Neo4j,"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.neo4j.driver.internal.messaging.v2;
-
-import org.neo4j.driver.internal.messaging.BoltProtocol;
-import org.neo4j.driver.internal.messaging.BoltProtocolVersion;
-import org.neo4j.driver.internal.messaging.MessageFormat;
-import org.neo4j.driver.internal.messaging.v1.BoltProtocolV1;
-
-public class BoltProtocolV2 extends BoltProtocolV1
-{
-    public static final BoltProtocolVersion VERSION = new BoltProtocolVersion( 2, 0 );
-
-    public static final BoltProtocol INSTANCE = new BoltProtocolV2();
-
-    @Override
-    public MessageFormat createMessageFormat()
-    {
-        return new MessageFormatV2();
-    }
-
-    @Override
-    public BoltProtocolVersion version()
-    {
-        return VERSION;
-    }
-}
