--- driver/src/main/java/org/neo4j/driver/internal/messaging/v2/MessageFormatV2.java
@@ -1,63 +1,0 @@
-/*
- * Copyright (c) 2002-2020 "Neo4j,"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.neo4j.driver.internal.messaging.v2;
-
-import org.neo4j.driver.internal.messaging.v1.MessageFormatV1;
-import org.neo4j.driver.internal.packstream.PackInput;
-import org.neo4j.driver.internal.packstream.PackOutput;
-
-public class MessageFormatV2 extends MessageFormatV1
-{
-    public static final byte DATE = 'D';
-    public static final int DATE_STRUCT_SIZE = 1;
-
-    public static final byte TIME = 'T';
-    public static final int TIME_STRUCT_SIZE = 2;
-
-    public static final byte LOCAL_TIME = 't';
-    public static final int LOCAL_TIME_STRUCT_SIZE = 1;
-
-    public static final byte LOCAL_DATE_TIME = 'd';
-    public static final int LOCAL_DATE_TIME_STRUCT_SIZE = 2;
-
-    public static final byte DATE_TIME_WITH_ZONE_OFFSET = 'F';
-    public static final byte DATE_TIME_WITH_ZONE_ID = 'f';
-    public static final int DATE_TIME_STRUCT_SIZE = 3;
-
-    public static final byte DURATION = 'E';
-    public static final int DURATION_TIME_STRUCT_SIZE = 4;
-
-    public static final byte POINT_2D_STRUCT_TYPE = 'X';
-    public static final int POINT_2D_STRUCT_SIZE = 3;
-
-    public static final byte POINT_3D_STRUCT_TYPE = 'Y';
-    public static final int POINT_3D_STRUCT_SIZE = 4;
-
-    @Override
-    public Writer newWriter( PackOutput output )
-    {
-        return new MessageWriterV2( output );
-    }
-
-    @Override
-    public Reader newReader( PackInput input )
-    {
-        return new MessageReaderV2( input );
-    }
-}
