--- driver/src/main/java/org/neo4j/driver/internal/messaging/v2/MessageWriterV2.java
@@ -1,30 +1,0 @@
-/*
- * Copyright (c) 2002-2020 "Neo4j,"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.neo4j.driver.internal.messaging.v2;
-
-import org.neo4j.driver.internal.messaging.v1.MessageWriterV1;
-import org.neo4j.driver.internal.packstream.PackOutput;
-
-public class MessageWriterV2 extends MessageWriterV1
-{
-    public MessageWriterV2( PackOutput output )
-    {
-        super( new ValuePackerV2( output ) );
-    }
-}
