+++ driver/src/main/java/org/neo4j/driver/internal/messaging/v4/MessageFormatV4.java
@@ -19,7 +19,7 @@
 package org.neo4j.driver.internal.messaging.v4;
 
 import org.neo4j.driver.internal.messaging.MessageFormat;
-import org.neo4j.driver.internal.messaging.v2.MessageReaderV2;
+import org.neo4j.driver.internal.messaging.common.CommonMessageReader;
 import org.neo4j.driver.internal.packstream.PackInput;
 import org.neo4j.driver.internal.packstream.PackOutput;
 
@@ -34,6 +34,6 @@ public class MessageFormatV4 implements MessageFormat
     @Override
     public Reader newReader( PackInput input )
     {
-        return new MessageReaderV2( input );
+        return new CommonMessageReader( input );
     }
 }
