+++ driver/src/main/java/org/neo4j/driver/internal/messaging/v4/MessageWriterV4.java
@@ -18,37 +18,21 @@
  */
 package org.neo4j.driver.internal.messaging.v4;
 
-import java.util.Map;
-
 import org.neo4j.driver.internal.messaging.AbstractMessageWriter;
 import org.neo4j.driver.internal.messaging.MessageEncoder;
-import org.neo4j.driver.internal.messaging.encode.BeginMessageEncoder;
-import org.neo4j.driver.internal.messaging.encode.CommitMessageEncoder;
-import org.neo4j.driver.internal.messaging.encode.DiscardMessageEncoder;
-import org.neo4j.driver.internal.messaging.encode.GoodbyeMessageEncoder;
-import org.neo4j.driver.internal.messaging.encode.HelloMessageEncoder;
-import org.neo4j.driver.internal.messaging.encode.PullMessageEncoder;
-import org.neo4j.driver.internal.messaging.encode.ResetMessageEncoder;
-import org.neo4j.driver.internal.messaging.encode.RollbackMessageEncoder;
-import org.neo4j.driver.internal.messaging.encode.RunWithMetadataMessageEncoder;
-import org.neo4j.driver.internal.messaging.request.BeginMessage;
-import org.neo4j.driver.internal.messaging.request.CommitMessage;
-import org.neo4j.driver.internal.messaging.request.DiscardMessage;
-import org.neo4j.driver.internal.messaging.request.GoodbyeMessage;
-import org.neo4j.driver.internal.messaging.request.HelloMessage;
-import org.neo4j.driver.internal.messaging.request.PullMessage;
-import org.neo4j.driver.internal.messaging.request.ResetMessage;
-import org.neo4j.driver.internal.messaging.request.RollbackMessage;
-import org.neo4j.driver.internal.messaging.request.RunWithMetadataMessage;
-import org.neo4j.driver.internal.messaging.v2.ValuePackerV2;
+import org.neo4j.driver.internal.messaging.common.CommonValuePacker;
+import org.neo4j.driver.internal.messaging.encode.*;
+import org.neo4j.driver.internal.messaging.request.*;
 import org.neo4j.driver.internal.packstream.PackOutput;
 import org.neo4j.driver.internal.util.Iterables;
+
+import java.util.Map;
 
 public class MessageWriterV4 extends AbstractMessageWriter
 {
     public MessageWriterV4( PackOutput output )
     {
-        super( new ValuePackerV2( output ), buildEncoders() );
+        super( new CommonValuePacker( output ), buildEncoders() );
     }
 
     private static Map<Byte,MessageEncoder> buildEncoders()
