+++ driver/src/test/java/org/neo4j/driver/internal/handlers/RunResponseHandlerTest.java
@@ -19,26 +19,17 @@
 package org.neo4j.driver.internal.handlers;
 
 import org.junit.jupiter.api.Test;
+import org.neo4j.driver.internal.messaging.v3.BoltProtocolV3;
+import org.neo4j.driver.internal.util.MetadataExtractor;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
-import org.neo4j.driver.internal.messaging.v1.BoltProtocolV1;
-import org.neo4j.driver.internal.messaging.v3.BoltProtocolV3;
-import org.neo4j.driver.internal.util.MetadataExtractor;
-
 import static java.util.Arrays.asList;
-import static java.util.Collections.emptyList;
-import static java.util.Collections.emptyMap;
-import static java.util.Collections.singletonMap;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static java.util.Collections.*;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.neo4j.driver.Values.value;
 import static org.neo4j.driver.Values.values;
 
@@ -115,11 +106,6 @@ class RunResponseHandlerTest
         assertEquals( keyIndex, handler.queryKeys().keyIndex() );
     }
 
-    @Test
-    void shouldReturnResultAvailableAfterWhenSucceededV1()
-    {
-        testResultAvailableAfterOnSuccess( "result_available_after", BoltProtocolV1.METADATA_EXTRACTOR );
-    }
 
     @Test
     void shouldReturnResultAvailableAfterWhenSucceededV3()
@@ -138,12 +124,12 @@ class RunResponseHandlerTest
 
     private static RunResponseHandler newHandler()
     {
-        return newHandler( BoltProtocolV1.METADATA_EXTRACTOR );
+        return newHandler( BoltProtocolV3.METADATA_EXTRACTOR );
     }
 
     private static RunResponseHandler newHandler( CompletableFuture<Throwable> runCompletedFuture )
     {
-        return newHandler( runCompletedFuture, BoltProtocolV1.METADATA_EXTRACTOR );
+        return newHandler( runCompletedFuture, BoltProtocolV3.METADATA_EXTRACTOR );
     }
 
     private static RunResponseHandler newHandler( MetadataExtractor metadataExtractor )
