+++ driver/src/test/java/org/neo4j/driver/internal/messaging/BoltProtocolTest.java
@@ -20,11 +20,8 @@ package org.neo4j.driver.internal.messaging;
 
 import io.netty.channel.embedded.EmbeddedChannel;
 import org.junit.jupiter.api.Test;
-
-import org.neo4j.driver.internal.messaging.v1.BoltProtocolV1;
-import org.neo4j.driver.internal.messaging.v2.BoltProtocolV2;
-import org.neo4j.driver.internal.messaging.v3.BoltProtocolV3;
 import org.neo4j.driver.exceptions.ClientException;
+import org.neo4j.driver.internal.messaging.v3.BoltProtocolV3;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
@@ -38,8 +35,6 @@ class BoltProtocolTest
     void shouldCreateProtocolForKnownVersions()
     {
         assertAll(
-                () -> assertThat( BoltProtocol.forVersion( BoltProtocolV1.VERSION ), instanceOf( BoltProtocolV1.class ) ),
-                () -> assertThat( BoltProtocol.forVersion( BoltProtocolV2.VERSION ), instanceOf( BoltProtocolV2.class ) ),
                 () -> assertThat( BoltProtocol.forVersion( BoltProtocolV3.VERSION ), instanceOf( BoltProtocolV3.class ) )
         );
     }
